Fix remaining tests by correcting paths and adding more assertions
