+++ test/testapps/basicdropwizardapp/.gitignore
