+++ test/testapps/basicdropwizardapp/pom.xml
