+++ test/testapps/basicdropwizardapp/src/main/java/com/github/sullis/dropwizard/helloworld/CustomJson.java
