+++ test/testapps/basicdropwizardapp/src/main/java/com/github/sullis/dropwizard/helloworld/HelloWorldServiceConfiguration.java
