+++ test/testapps/basicdropwizardapp/src/main/java/com/github/sullis/dropwizard/helloworld/api/HelloMessage.java
