+++ test/testapps/basicdropwizardapp/startappfromcode.sh
