+++ test/testapps/basicplayapp/build.sbt
