+++ test/testapps/basicplayapp/conf/application.conf
