+++ test/testapps/basicplayapp/project/build.properties
