+++ test/testapps/basicplayapp/project/plugins.sbt
