+++ test/testapps/openplayapp/project/build.properties
