+++ test/testapps/openplayapp/project/plugins.sbt
